Small tidying up from SearchPane -> LeftPane
